Fix failure capture arguments in Twig functions url() and path()
Octagon provides two functions that allow the developer to generate
URLs from inside a Twig template. The function allows the developer
to supply the URL generator with placeholder values in an array as
an argument. The URL generator properly generates the stored URL
string, but fails to place the placeholder values in the URL.

The failure to replace placeholders with the given placeholder
values was caused by two things:

First, I was treating the as options. Once I fixed that, I then
realized that the URLs still weren't being generated correctly.
It turned out that Twig stores the given args array in another
array. So, while I was accessing $argsArray[target], I should have
been accessing $argsArray[0][target].

I also reduced duplicate code by placing the reused code in a
function:

- extractVariableName(): Pulls placeholder variable name from
  {placeholder}

I also renamed subPatternArgs() to substitutePatternArgs() for
better clarity and more meaningfulness.